movie/requirements.txt
django==2.0.2
psycopg2==2.8.3
djangorestframework==3.7.7
django-debug-toolbar==2.0
django-extensions==1.9.9
pillow==5.0.0
djangorestframework-filters==0.10.2
django-filter==1.1.0
django-cors-headers==2.2.0
django-phonenumber-field==2.0.0